texassolver-web/api/app/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.6.1
aiofiles==24.1.0
python-multipart==0.0.12

# Database
sqlalchemy==2.0.36
asyncpg==0.29.0
alembic==1.14.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
